requirements.txt
boto
boto3
botocore
configparser
coverage
Django
django-allauth
django-appconf
django-boto
django-ckeditor
django-compressor
django-jalali
django-jet
django-secure
django-storages
djangorestframework
docutils
enum-compat
enum34
flake8
futures
gevent
greenlet
gunicorn
html5lib
jdatetime
jmespath
jsonfield
libwapiti
mccabe
nltk
oauthlib
Pillow
psycopg2
pycodestyle
pyflakes
python-dateutil
python-openid
pytz
rcssmin
requests
requests-oauthlib
rjsmin
s3transfer
six
